Removes move.admin from group ID. Renames source code package.
