+++ sr-proxy/pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
     <modelVersion>4.0.0</modelVersion>
-    <groupId>no.difi.move.admin</groupId>
+    <groupId>no.difi.move.serviceregistry</groupId>
     <artifactId>sr-proxy</artifactId>
     <version>1.0.4-SNAPSHOT</version>
     <packaging>jar</packaging>
