+++ sr-proxy/src/main/java/no/difi/move/serviceregistry/auth/OidcTokenResponse.java
@@ -1,0 +1,49 @@
+package no.difi.move.serviceregistry.auth;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategy;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import com.google.common.base.MoreObjects;
+
+@JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
+public class OidcTokenResponse {
+
+    private String accessToken;
+    private Integer expiresIn;
+    private String scope;
+
+    public OidcTokenResponse() {
+    }
+
+    public String getAccessToken() {
+        return accessToken;
+    }
+
+    public void setAccessToken(String accessToken) {
+        this.accessToken = accessToken;
+    }
+
+    public int getExpiresIn() {
+        return expiresIn;
+    }
+
+    public void setExpiresIn(Integer expiresIn) {
+        this.expiresIn = expiresIn;
+    }
+
+    public String getScope() {
+        return scope;
+    }
+
+    public void setScope(String scope) {
+        this.scope = scope;
+    }
+
+    @Override
+    public String toString() {
+        return MoreObjects.toStringHelper(this)
+                .add("accessToken", accessToken)
+                .add("expiresIn", expiresIn)
+                .add("scope", scope)
+                .toString();
+    }
+}
