--- sr-proxy/src/main/java/no/difi/serviceregistry/proxy/auth/OidcErrorHandler.java
@@ -1,35 +1,0 @@
-package no.difi.serviceregistry.proxy.auth;
-
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.io.IOUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.client.ClientHttpResponse;
-import org.springframework.web.client.DefaultResponseErrorHandler;
-import org.springframework.web.client.HttpClientErrorException;
-import org.springframework.web.client.HttpServerErrorException;
-import org.springframework.web.client.RestClientException;
-
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-
-@Slf4j
-class OidcErrorHandler extends DefaultResponseErrorHandler {
-
-    @Override
-    public void handleError(ClientHttpResponse response) throws IOException {
-        log.error("Response from OIDC: " + IOUtils.toString(response.getBody(), StandardCharsets.UTF_8));
-        HttpStatus statusCode = getHttpStatusCode(response);
-        switch (statusCode.series()) {
-            case CLIENT_ERROR:
-                throw new HttpClientErrorException(statusCode, response.getStatusText(),
-                        response.getHeaders(), getResponseBody(response), getCharset(response));
-            case SERVER_ERROR:
-                throw new HttpServerErrorException(statusCode, response.getStatusText(),
-                        response.getHeaders(), getResponseBody(response), getCharset(response));
-            default:
-                throw new RestClientException("Unknown status code [" + statusCode + "]");
-        }
-    }
-}
